Level 4 and perspective try
I added a new fourth level in the game that you can now access upon finishing the level 3. If this new level i want to try to implement more dynamics and moving mechanics (such as moving walls or on/off bridge).
Right now, the level don't have any real mechanics. But i tried to improve my (really bad) skills in drawing so i added sort of perspective to the level. You can now see the ocean at the bottom of the structure the player is in.

The textures are not really good but i will improve them when i continue working on the level.