New movement mechanic
I successfully added a new movement mechanic to the game. Now, instead of instantly going to the next tile, the enemies will slide to it in the same way as the player is doing. After each move, the player will now have a little cooldown as the enemies are moving. This mechanic is more close to what I initially thought for the game, kinda like a chess turn based movement game.

However, this new mechanic make the player slower. In order to have a good experience, i am gonna remap the differents levels so that you don't need to click a hundred time to finish one. For now, the first level is only working, i will implement the new movement mechanic to the others later.