Additional keybinds & movement correction
I added the arrow to the input that you use for moving. In addition to that, I changed how the game check the inputs, so enemy units can now move better.

I also made a little art addition to the title screen (I find it terrible, i'm definitly not an artist)