Ressources and saves files
I changed the hierarchy of the project in order to store the ressources used by the buttons and others objects such as TileMapLayers (since it's always the same ressource, it was a waste of time and energy to create a new ressource each time).

I also updated how the save system work. Now, the play button will always put you at the last level you were doing. There is now options if you want to reset your progression.
(N.B. The save system don't seem to work in the web version, i will fix it if i find a solution).